Refactor telemetry handlers and update logMetrics
Moved telemetry handler classes from TelemetryFactory.swift to a new TelemetryFactory.Handlers.swift file and implemented metric logging in CKCounter and CKFloatingCounter. Updated logMetrics to use Telemetry.Export instead of Telemetry and clarified documentation for telemetry parameter.
