+++ Sources/Scout/Core/TelemetryFactory.swift
@@ -5,6 +5,7 @@
 // license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import Foundation
 import Metrics
 
 struct TelemetryFactory: MetricsFactory {
@@ -47,60 +48,4 @@ struct TelemetryFactory: MetricsFactory {
     func destroyTimer(_ handler: TimerHandler) {
 
     }
-}
-
-final class CKCounter: CounterHandler {
-    func increment(by: Int64) {
-
-    }
-
-    func reset() {
-
-    }
-}
-
-final class CKFloatingCounter: FloatingPointCounterHandler {
-    func increment(by: Double) {
-
-    }
-
-    func reset() {
-
-    }
-}
-
-final class CKMeter: MeterHandler {
-    func set(_ value: Int64) {
-
-    }
-
-    func set(_ value: Double) {
-
-    }
-
-    func increment(by: Double) {
-
-    }
-
-    func decrement(by: Double) {
-
-    }
-}
-
-final class CKRecorder: RecorderHandler {
-    func record(_ value: Int64) {
-
-    }
-
-    func record(_ value: Double) {
-
-    }
-
-}
-
-final class CKTimer: TimerHandler {
-    func recordNanoseconds(_ duration: Int64) {
-
-    }
-
 }
